Add files via upload
As described in our paper for SIGIR 2023 submission, we design a questionnaire to acquire user requirements on open-source recommendation library.
We distribute questionnaires mainly for researchers in the field of recommender systems, and collect 153 valid (without missing slots) feedbacks in total. 
This is the original data of the collected questionnaires.